requirements.txt
Flask==3.0.0
rcpchgrowth==4.3.8
python-dateutil==2.8.2
gunicorn==21.2.0
Flask-Limiter==3.5.0
pytest==7.4.3
scipy==1.11.4
